feat(cli): refactor pipes components
