+++ screenpipe-js/cli/src/commands/index.ts
@@ -1,8 +1,5 @@
-export { createPipeCommand } from './create';
-export { appCommands } from './app';
 export { loginCommand } from './login';
 export { logoutCommand } from './logout';
-export { publishCommand } from './publish';
-export { registerCommand } from './register';
-export { listVersionsCommand } from './list-versions';
+export { pipeCommands } from "./pipe"
+export { appCommands } from './app';
 export { componentsCommands } from "./components"
