+++ screenpipe-js/cli/src/commands/pipe/create.ts
